feat: remove WalletUiSize property
This commit removes the WalletUiSize property from the @wallet-ui/react components.

The goal is to make this a headless library and as such it should not care these kind of things.

Instead we might want to add more places where people can pass through a `classNames` prop to make ad-hoc tweaks.